+++ main_base.py
@@ -140,6 +140,7 @@ def main(
     verbose: bool = True,
     test_mode: Literal["on", "off", "mock"] = "on",
     max_workers: int | None = None,
+    **kwargs
 ) -> dict:
     resources: GraphResources | None = None
     try:
@@ -223,16 +224,17 @@ def main(
         return placeholders_output
 
     finally:
-        qdrant_service = getattr(resources, "qdrant_service", None)
-        if (
-            qdrant_service
-            and qdrant_service.client.collection_exists(collection_name)
-            and is_valid_uuid4_hex(collection_name)
-        ):
-            qdrant_service.client.delete_collection(collection_name)
-            logger.info(
-                f"collection <{collection_name}> name is valid uuid and was deleted"
-            )
+        if "save_db" not in kwargs:
+            qdrant_service = getattr(resources, "qdrant_service", None)
+            if (
+                qdrant_service
+                and qdrant_service.client.collection_exists(collection_name)
+                and is_valid_uuid4_hex(collection_name)
+            ):
+                qdrant_service.client.delete_collection(collection_name)
+                logger.info(
+                    f"collection <{collection_name}> name is valid uuid and was deleted"
+                )
 
 
 if __name__ == "__main__":
